Bring the Qwen3.5 text decoder up on device at TP=8
Adds the Phase 6 tooling that was never built and runs the sharded model
through the real vLLM engine on two Trainium devices. TP=8 works: eight
ranks, sharded GDN projections, sequence-parallel collectives, the hybrid
Mamba/attention state cache, and both prefill and decode compiles, all
generating tokens end to end.

The tiny checkpoint is generated from HuggingFace's own Qwen3_5TextModel so
every key and shape is right by construction, and its geometry puts TP=8 in
the same per-rank regime the 27B reaches at TP=16 (1 key head, 3 value
heads).

Three dtype bugs had to be fixed to get there. All were invisible to the
test suite because CPU torch silently promotes; only the device rejects
them. The first two predate the tensor-parallel work:

  - The GDN built all six projections at torch's fp32 default instead of
    config.torch_dtype, leaving 48 of the real model's 64 layers at double
    width. MLP and Attention both pass the dtype; the GDN was the outlier.
  - One fp32 validity mask was applied to both paged states, silently
    promoting the bf16 conv state, which the depthwise conv then rejected
    with "nc_matmul: if one input is tfloat32/float32, both must be".
  - The conv now casts its state to the activation dtype at the kernel
    boundary, so the call is dtype-consistent however the cache is typed.

get_kv_spec now declares both GDN states fp32. The runner carves both from
one page and the model writes both back in place, and aot_autograd rejects
mutating two differing-dtype views of one storage. Matching them costs the
real 27B 1.9% state memory; giving each state its own allocation is the
better fix and is left as a deliberate choice.

The chunk-scan kernel still fails neuronx-cc codegen on the decode graph
(NCC_IXGM002), so that one kernel runs on its torch fallback. Attribution
is from a full 2x2 of device runs; the conv kernel is not implicated, and
the LNC=2 split was tested and ruled out.

docs/model-dev/qwen3-5-tp8-device-bringup.md records all of it, including
two blockers that stop a real checkpoint before any sharding runs: the
factory signature against the multimodal branch, and hybrid page-size
misalignment that the 27B hits at every block size.

+++ vllm_neuron/model/qwen3_5/nki_gdn.py
@@ -119,7 +119,16 @@ def causal_conv1d_with_state(
     seq_len = hidden_states.shape[-1]
     channels = hidden_states.shape[1]
 
-    extended = torch.cat([conv_state, hidden_states], dim=-1)
+    # The conv runs at the activation dtype, so the state joins it there. The
+    # cached state may be stored wider -- it shares a page with the fp32
+    # recurrent accumulator, and both must carry the same dtype for the
+    # backend to allow writing them back in place (see model.py::get_kv_spec).
+    # Concatenating without this cast promotes the whole window to fp32 and the
+    # kernel then rejects the pair: "nc_matmul: if one input is
+    # tfloat32/float32, both must be. Got stationary=bfloat16, moving=float32".
+    # This function already treats hidden_states.dtype as authoritative -- it
+    # casts the result back to it on return.
+    extended = torch.cat([conv_state.to(hidden_states.dtype), hidden_states], dim=-1)
 
     if _can_use_conv_kernel(hidden_states, channels):
         # nkilib wants [N, C, 1, W] and [C, 1, 1, S]; the state supplies the
